backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.2
sentence-transformers==2.7.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
spacy==3.7.0
dateparser==1.1.8
torch==2.1.0
ultralytics==8.3.0
transformers==4.36.0
openai==1.12.0
httpx==0.24.1
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0

